requirements.txt
bs4
html5lib
Mastodon.py
requests
psycopg2
